requirements.txt
# Core dependencies
numpy>=1.24.0
cryptography>=41.0.0
pyyaml>=6.0

# CLI
click>=8.1.0

# AWS Braket (optional - for real quantum devices)
amazon-braket-sdk>=1.50.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

